# Improve `write.GSD` and `get_snapshot` performance.