<!-- Please confirm that your work is based on the correct branch. -->
<!-- Bug fixes should be based on *trunk-patch*. -->
<!-- Backwards compatible new features should be based on *trunk-minor*. -->
<!-- Incompatible API changes should be based on *trunk-major*. -->

## Description

<!-- Describe your changes in detail. -->
* Rewrite GSD writer to directly use local particle properties in place of `takeSnapshot`.
* Rewrite serial `takeSnapshot` code path to improve performance.
* The MPI code path in `takeSnapshot` remains slow for now, but could later be rewritten with code similar to the new code in `GSDDumpWriter`.
* Topology (bond, angle, ...) snapshots also remain slow (most use cases only write this in the first frame).
* Remove `ExternalCallback` code, as it is unused and relies on the now removed `std::map` in the snapshot.
* Expand the `dynamic` flag to take in a list of individual particle fields. Keep the existing `'property'`, `'attribute'`, and `'momentum'` meta-fields for backward compatibility.

## Motivation and context

<!--- Why is this change required? What problem does it solve? -->
`write.GSD` is extremely slow in current releases. On profiling, I found the majority of the time was spent in creating and iterating through the `std::map` used by `takeSnapshot`. Next after that is std::vector allocations in the snapshot objects.

* The new GSD writer accesses only the particle data needed given the user parameters (e.g. in the default dynamic=['property'], only position and orientation are accessed). 
* GSD now only sorts the non-default values it will write. 
* GSD now holds on to the `std::vector` instances from one frame to the next to avoid reallocation costs.
* In MPI, GSD now uses an optimized k-way merge sort to order the particles in increasing tag order.

The optimizations to `takeSnapshot` are in this pull request because I took those steps first in an effort to boost GSD performance. There is more work to make `takeSnapshot` as fast in MPI as GSD is, but I leave that work for a future pull request.

In log only GSD files (`filter=hoomd.filter.Null()`), performance is now ~1000x faster (speedup depends on system size) as GSD no longer accesses any particle data in this case - before it always called `takeSnapshot`.

## How has this been tested?

<!--- Please describe in detail how you tested your changes. -->
Existing unit tests pass.

Will post benchmark results when ready to review.

<!--- Please build the sphinx documentation and check that any changes to
      documentation display properly. -->

## Change log

<!-- Propose a change log entry. -->
```
Changed:

* Increase performance of ``hoomd.write.GSD``.
* Increase performance of ``hoomd.State.get_snapshot`` in serial.
* `hoomd.write.GSD.dynamic` now allows fine grained control over individual particle fields.
```

## Checklist:

- [x] I have reviewed the [**Contributor Guidelines**](https://github.com/glotzerlab/hoomd-blue/blob/trunk-minor/CONTRIBUTING.rst).
- [x] I agree with the terms of the [**HOOMD-blue Contributor Agreement**](https://github.com/glotzerlab/hoomd-blue/blob/trunk-minor/ContributorAgreement.md).
- [x] My name is on the list of contributors (`sphinx-doc/credits.rst`) in the pull request source branch.