@joaander 2023-04-19
Benchmark results on `cheme-hodges` (NVME SSD):
* 1 rank, N=64000
  * This branch: 241 MiB/s
  * `trunk-major`: 34.8 MiB/s
* 1 rank, N=1000000
  * This branch: 203 MiB/s
  * `trunk-major`: 33.2 MiB/s
* 6 ranks, N=64000:
  * This branch: 150 MiB/s
  * `trunk-major`: 13.1 MiB/s
* 6 ranks, N=1000000)
  * This branch: 191 MiB/s
  * `trunk-major`: 11.1 MiB/s

I will add this benchmark to `hoomd-benchmarks` soon. To test, I ran` python3 -m hoomd_benchmarks.write_gsd --device CPU -v --repeat=5 --warmup_steps=100 --benchmark_steps=500 -N <N> --bandwidth`.