@joaander 2023-04-20
Benchmark results on `summit` (Alpine):
* 1 rank, N=64000
  * This branch: 12.2 MiB/s (16.5 frames/second)
  * `trunk-major`: 8.7 MiB/s (11.6 frames/second)
* 1 rank, N=1000000
  * This branch:  84.4 MiB/s (7.5 frames/second)
  * `trunk-major`: 14.0 MiB/s (1.28 frames/second)
* 6 ranks, N=64000:
  * This branch: 11.5 MiB/s
  * `trunk-major`:  4.79 MiB/s
* 6 ranks, N=1000000)
  * This branch: 81.6 MiB/s
  * `trunk-major`:  6.02 MiB/s