@joaander 2023-04-20
With the new fine grained dynamic control, setting `dynamic=['particles/position']` offers a slight performance increase over the default `dynamic=['property']` (which includes orientation).