@joaander 2023-04-20
@b-butler @tommy-waltmann I ask you both to review as this is a critical piece of code. Exiting unit test pass. Adding new unit tests that cover all 1024+ combinations of parameters is not feasible.