@joaander 2023-04-24
And on Frontier (Orion):
* 1 rank, N=64000
  * This branch: 11.64 MiB/s (15.5 frames/s)
* 1 rank, N=1000000
  * This branch:  99.8 MiB/s
* 6 ranks, N=64000:
  * This branch: 10.75 MiB/s (8.5 frames/s)
* 6 ranks, N=1000000)
  * This branch: 101 MiB/s

I have confirmed that the per-frame `fsync` for stability limits the performance on these massive scale systems to ~10-15 frames/s. When I remove the `fsync`, the benchmark achieves ~500-600 MiB/s regardless of system size.